# orchestratorOpenApi.json
{
    "openapi": "3.0.0",
    "servers": [
        {
            "url": "/api/v1/service"
        }
    ],
    "info": {
        "title": "Orchestrator Spec File",
        "version": "0.1.9"
    },
    "paths": {
...
        "/login": {
            "post": {
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "username",
                                    "password",
                                    "organization"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "organization": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Just a description"
                    }
                }
            }
        },
        "/logout": {
            "delete": {
                "parameters": [
                    {
                        "in": "header",
                        "name": "X-Auth-Token",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Logged out successfully."
                    }
                }
            }
        },
...
    }
}